Flytt securelog til feature - lokalitet
