--- felles/server/src/main/java/no/nav/vedtak/server/rest/RestSikkerLoggFeature.java
@@ -1,24 +1,0 @@
-package no.nav.vedtak.server.rest;
-
-import jakarta.ws.rs.core.Feature;
-import jakarta.ws.rs.core.FeatureContext;
-
-
-public class RestSikkerLoggFeature implements Feature {
-
-    private static boolean sikkerloggEnabled = false;
-
-    @Override
-    public boolean configure(final FeatureContext context) {
-        setSikkerloggEnabled();
-        return true;
-    }
-
-    public static boolean erSikkerloggEnabled() {
-        return sikkerloggEnabled;
-    }
-
-    private static void setSikkerloggEnabled() {
-        RestSikkerLoggFeature.sikkerloggEnabled = true;
-    }
-}
